Issue #48, app.py result_dao.py refactored and EXEC logic moved into db_connection.py
