+++ src/db_connection.py
@@ -1,7 +1,11 @@
 import os
 import pyodbc
 from dotenv import load_dotenv
+import datetime
+import sqlparse
+
 import src.db_DAO as db_dao
+import src.models as models
 
 load_dotenv()
 
@@ -169,3 +173,150 @@ def get_actual_db_name():
     if db_connection:
         return db_connection.db_name
     return None
+
+
+def safe_pretty_json(record: dict) -> dict:
+    """Convert record values to safe JSON-serializable format"""
+    safe_record = {}
+    for k, v in record.items():
+        if k == "Query Text":
+            safe_record[k] = sqlparse.format(v, keyword_case='upper', output_format='sql', reindent=True)
+        elif isinstance(v, datetime.datetime):
+            safe_record[k] = v.isoformat()
+        elif isinstance(v, datetime.date):
+            safe_record[k] = v.isoformat()
+        elif isinstance(v, datetime.time):
+            safe_record[k] = v.isoformat()
+        elif isinstance(v, bytes):
+            # Convert bytes to hex string for display
+            safe_record[k] = v.hex() if v else ''
+        else:
+            safe_record[k] = v
+    return safe_record
+
+def exec_more_info(record, index_records, finding_records):
+    with get_connection() as sql_server_conn:
+        cursor = sql_server_conn.cursor()
+        cursor.execute(record.more_info)
+
+            # Skip to the result set with index data
+        while cursor.description is None:
+            if not cursor.nextset():
+                break
+
+            # Process first result set (Q1 - Index details)
+        if cursor.description:
+            columns = [desc[0] for desc in cursor.description]
+            rows = cursor.fetchall()
+
+                # print(f"Fetched {len(rows)} rows from sp_BlitzIndex output")
+                # Convert rows to list of DBIndexRecord objects and skip first row (Q1)
+            for i, row in enumerate(rows):
+                if i == 0:  # Skip first row (Q1)
+                    continue
+                row_dict = dict(zip(columns, row))
+                    # Serialize the data for storage
+                serialized_row = safe_pretty_json(row_dict)
+
+                    # Map sp_BlitzIndex columns to DBIndexRecord fields
+                mapped_data = {}
+
+                    # Map the columns from sp_BlitzIndex to DBIndexRecord fields
+                column_mapping = {
+                        'Details: db_schema.table.index(indexid)': 'db_schema_object_indexid',
+                        'Definition: [Property] ColumnName {datatype maxbytes}': 'index_definition',
+                        'Secret Columns': 'secret_columns',
+                        'Fillfactor': 'fill_factor',
+                        'Usage Stats': 'index_usage_summary',
+                        'Op Stats': 'index_op_stats',
+                        'Size': 'index_size_summary',
+                        'Compression Type': 'partition_compression_detail',
+                        'Lock Waits': 'index_lock_wait_summary',
+                        'Referenced by FK?': 'is_referenced_by_foreign_key',
+                        'FK Covered by Index?': 'fks_covered_by_index',
+                        'Last User Seek': 'last_user_seek',
+                        'Last User Scan': 'last_user_scan',
+                        'Last User Lookup': 'last_user_lookup',
+                        'Last User Write': 'last_user_update',
+                        'Created': 'create_date',
+                        'Last Modified': 'modify_date',
+                        'Page Latch Wait Count': 'page_latch_wait_count',
+                        'Page Latch Wait Time (D:H:M:S)': 'page_latch_wait_time',
+                        'Page IO Latch Wait Count': 'page_io_latch_wait_count',
+                        'Page IO Latch Wait Time (D:H:M:S)': 'page_io_latch_wait_time',
+                        'Create TSQL': 'create_tsql',
+                        'Drop TSQL': 'drop_tsql'
+                    }
+
+                for sp_column, db_field in column_mapping.items():
+                    if sp_column in serialized_row:
+                        value = serialized_row[sp_column]
+                            # Convert boolean strings to integers for FK fields
+                        if db_field == 'is_referenced_by_foreign_key' and isinstance(value, str):
+                            mapped_data[db_field] = 1 if value.lower() == 'true' else 0
+                        else:
+                            mapped_data[db_field] = value
+
+                    # Create DBIndexRecord object with mapped data
+                index_record = models.DBIndexRecord(pbi_id=record.pbi_id, **mapped_data)
+                index_records.append(index_record)
+
+            # Process second result set (Q2 - Missing index findings)
+        if cursor.nextset() and cursor.description:
+            columns = [desc[0] for desc in cursor.description]
+            rows = cursor.fetchall()
+
+                # Convert rows to list of DBFindingRecord objects
+            for row in rows:
+                row_dict = dict(zip(columns, row))
+
+                    # Map columns from Q2 to DBFindingRecord fields
+                finding_data = {
+                        'pbi_id': record.pbi_id,
+                        'finding': row_dict.get('Finding'),
+                        'url': row_dict.get('URL'),
+                        'estimated_benefit': row_dict.get('Estimated Benefit'),
+                        'missing_index_request': row_dict.get('Missing Index Request'),
+                        'estimated_impact': row_dict.get('Estimated Impact'),
+                        'create_tsql': row_dict.get('Create TSQL'),
+                        'sample_query_plan': row_dict.get('Sample Query Plan')
+                    }
+
+                    # Create DBFindingRecord object
+                finding_record = models.DBFindingRecord(**finding_data)
+                finding_records.append(finding_record)
+
+
+
+def exec_blitz(procedure_name):
+    records = []
+
+    # Get database name for procedures that require it
+    db_connection = db_dao.get_db(get_actual_db_id())
+    database_name = db_connection.db_name if db_connection else None
+
+    with get_connection() as db_connection:
+        cursor = db_connection.cursor()
+
+            # Add @DatabaseName parameter for sp_BlitzCache and sp_BlitzIndex
+        if procedure_name == 'sp_BlitzCache' and database_name:
+            cursor.execute(f"EXEC {procedure_name} @DatabaseName = ?", (database_name,))
+        elif procedure_name == 'sp_BlitzIndex' and database_name:
+            cursor.execute(f"EXEC {procedure_name} @IncludeInactiveIndexes=1, @Mode=4, @DatabaseName = ?", (database_name,))
+        else:
+            cursor.execute(f"EXEC {procedure_name}")
+
+        while cursor.description is None:
+            if not cursor.nextset():
+                return False, []
+
+        columns = [desc[0] for desc in cursor.description]
+        rows = cursor.fetchall()
+        records = []
+        for row in rows:
+            full = dict(zip(columns, row))
+                # Apply serialization for storage
+            serialized_full = safe_pretty_json(full)
+                # Store the original SQL Server data for the DAO
+            records.append(serialized_full)
+    return True, records
